requirements.txt
Pillow
pytorch
matplotlib
transformers==4.21.2